Typescriptize SiteUrlSettings
